added the functionality to generate the paths for the pagination
