+++ config/index.js
@@ -1,0 +1,1 @@
+export const POSTS_PER_PAGE = 3;
